io: move decompression code to separate file
