+++ src/io/mod.rs
@@ -12,16 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use anyhow::{bail, Context, Result};
+use anyhow::{bail, Result};
 use bincode::Options;
-use flate2::read::GzDecoder;
-use std::io::{self, BufRead, ErrorKind, Read, Write};
-use std::result;
-use xz2::read::XzDecoder;
+use std::io::{ErrorKind, Read, Write};
 
+mod compress;
 mod hash;
 mod limit;
 
+pub use self::compress::*;
 pub use self::hash::*;
 pub use self::limit::*;
 
@@ -81,43 +80,6 @@ pub fn copy_exactly_n(
         );
     }
     Ok(n)
-}
-
-enum CompressDecoder<R: BufRead> {
-    Uncompressed(R),
-    Gzip(GzDecoder<R>),
-    Xz(XzDecoder<R>),
-}
-
-pub struct DecompressReader<R: BufRead> {
-    decoder: CompressDecoder<R>,
-}
-
-/// Format-sniffing decompressor
-impl<R: BufRead> DecompressReader<R> {
-    pub fn new(mut source: R) -> Result<Self> {
-        use CompressDecoder::*;
-        let sniff = source.fill_buf().context("sniffing input")?;
-        let decoder = if sniff.len() > 2 && &sniff[0..2] == b"\x1f\x8b" {
-            Gzip(GzDecoder::new(source))
-        } else if sniff.len() > 6 && &sniff[0..6] == b"\xfd7zXZ\x00" {
-            Xz(XzDecoder::new(source))
-        } else {
-            Uncompressed(source)
-        };
-        Ok(Self { decoder })
-    }
-}
-
-impl<R: BufRead> Read for DecompressReader<R> {
-    fn read(&mut self, buf: &mut [u8]) -> result::Result<usize, io::Error> {
-        use CompressDecoder::*;
-        match &mut self.decoder {
-            Uncompressed(d) => d.read(buf),
-            Gzip(d) => d.read(buf),
-            Xz(d) => d.read(buf),
-        }
-    }
 }
 
 /// Provides uniform bincode options for all our serialization operations.
